Backend-Core/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
email_validator==2.2.0
motor==3.3.1
beanie==1.23.3
pymongo==4.7.0